Added an additional option called :highlighted_date_check. This is a visual cue; non-highlighted dates are still choosable.
Added a new rule (highlighted) on the css files, used for this change.

Completed the comments on form_helpers by including a mention of :valid_date_check and :highlighted_date_check.

Added one unitary case, but I'm not sure it works ok... please review.

Examples of use:

This highlights halloween (31st of October):
  <%= calendar_date_select_tag "near_halloween_day", "", :highlighted_date_check => "date.getDate() == 31 && date.getMonth() == 10" %>
This highlights fridays:
  <%= calendar_date_select_tag "casual_day", "", :highlighted_date_check => "date.getDay() == 5" %>